requirements.txt
asgiref==3.11.1
Django==6.0.3
django-axes==8.3.1
django-colorfield==0.14.0
django-storages==1.14.6
django-tinymce==5.0.0
nh3==0.3.4
pillow==12.1.1
python-dotenv==1.2.2
sqlparse==0.5.5
